+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Codicon Inspector extension will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added a **Codicons** status bar button that opens the Codicon Inspector panel.
+
 ## [0.0.8] - 2026-04-23
 
 ### Added